feat: add geographic grounding to Research with AI
Extract PostGIS boundary lookup from serperService into shared
geoService module, then wire it into geminiService's multi-pass
research. Research prompts now include geographic context like
"Located in Hampton Hills Park, Summit County, Akron, OH" — the
same grounding that improved news/events search relevance by 80-100%.
